homework-3 split table component file to individual components files
