+++ src/app/components/table/row-cell.tsx
@@ -1,0 +1,11 @@
+import React, { ReactNode } from 'react';
+
+type RowCellProps = {
+   children?: ReactNode;
+}
+
+const RowCell = ({children}:RowCellProps) => (
+   <td className="p-[16px]">{children}</td>
+);
+
+export default RowCell;
